Bioconductor/R-packages
This is the file to create a singularity container on the HPC to be used in the pipeline. This will let you use R, and in R the packages DESeq2, Rsubread, and EnhancedVolcano will be installed also. Instructions on how to build the container from this file will be in the readme. This helps with reproducibility.